Download imagery directly from Library of Congress (#33)
* no more than four unnamed params

* LoC download script

* clean up

* retry 520, resume text

* test for loc.iiif.json

+++ fit.sh
@@ -13,11 +13,23 @@ uv run mapsnap/georef_from_labels.py $dir/*.scaled.jpg \
     --min-short-side 20 \
     --min-confidence 0.15
 
+# --edge-margin 0
+
+if [ -e $dir/main.iiif.json ]; then
+    ref_iiif=$dir/main.iiif.json
+else
+    ref_iiif=$dir/loc.iiif.json
+fi
+
 uv run python mapsnap/make_iiif_georef.py \
-    $dir/main.iiif.json $dir'/*.georef.json' \
+    $ref_iiif $dir'/*.georef.json' \
     --centerlines $dir/centerlines.geojson \
     --output $dir/$tag.iiif.json
 
-uv run python mapsnap/compare_iiif_georef.py \
-    $dir/main.iiif.json $dir/$tag.iiif.json \
-    | tee $dir/$tag.txt
+if [ -e $dir/main.iiif.json ]; then
+    uv run python mapsnap/compare_iiif_georef.py \
+        $dir/main.iiif.json $dir/$tag.iiif.json \
+        | tee $dir/$tag.txt
+else
+    echo "\nNo main.iiif.json, skipping comparison step.\n"
+fi
